requirements.txt
scrapy 
beautifulsoup4 
pymongo 
python-decouple